+++ packages/@dicebear/big-smile/src/components/index.ts
@@ -1,0 +1,5 @@
+export { face } from './face';
+export { mouth } from './mouth';
+export { eyes } from './eyes';
+export { hair } from './hair';
+export { accessories } from './accessories';