+++ packages/@dicebear/big-smile/src/index.ts
@@ -1,0 +1,21 @@
+/*!
+ * Custom Avatar (@dicebear/big-smile)
+ *
+ * Code licensed under MIT License.
+ * Copyright (c) 2021 Florian Körner
+ *
+ * Design "Custom Avatar" by Ashley Seo licensed under CC BY 4.0.
+ * Source: https://www.figma.com/community/file/881358461963645496
+ * License: https://creativecommons.org/licenses/by/4.0/
+ */
+
+import { utils } from '@dicebear/avatars';
+import { style } from './core';
+
+let { create, meta, schema } = style;
+
+export { create, meta, schema };
+export { Options } from './options';
+
+/** @deprecated will be removed in Version 5.0 */
+export default utils.style.createLegacyWrapper(style);