+++ packages/@dicebear/collection/src/index.ts
@@ -15,3 +15,4 @@ export * as pixelArtNeutral from '@dicebear/pixel-art-neutral';
 export * as openPeeps from '@dicebear/open-peeps';
 export * as personas from '@dicebear/personas';
 export * as miniavs from '@dicebear/miniavs';
+export * as bigSmile from '@dicebear/big-smile';