+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `@dicebear/personas` New Avatar style by draftbit.com.
 - `@dicebear/miniavs` New Avatar style by Webpixels.
+- `@dicebear/big-smile` New Avatar style by Ashley Seo.
 
 ### Changed